Remove dependence on pstopdf
pstopdf is no longer included in the ConTeXt package in modern Ubuntu.
One option is ps2pdf, but it is not a drop-in replacement.  Instead,
just have latex pull in the eps files directly, which works fine and
gets rid of the pstopdf requirement.

Also remove the gnuplot generated .tex files from the repository to
clear up any confusion.  Delete them on 'make clean'.