Support flagging events as holidays, grey them out
Add a boolean flag for if an event is a holiday.  If it is a holiday,
grey it out on the small month views

The holiday days are not greyed out on the larger month tables or the
weekly planner, because those effects might be confusing with greyed out
days from previous/future months or just not look good, and because the
event text is already displayed on those views.